requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
bleach==6.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==3.2.20
django-classy-tags==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.14.0
django-nyt==1.3
django-oauth-toolkit==2.3.0
django-sekizai==4.1.0
djangochannelsrestframework==0.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
environ==1.0
h11==0.12.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
hypercorn==0.12.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jwcrypto==1.5.0
jwt==1.3.1
kombu==5.3.4
loguru==0.7.0
Markdown==3.3.7
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.22
Pillow==10.0.0
priority==2.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tinycss2==1.2.1
toml==0.10.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
vonage==3.11.0
vonage-jwt==1.1.0
wcwidth==0.2.6
webencodings==0.5.1
wiki==0.10
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.2.0
zope.interface==6.1
